backend/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
pydantic
python-dotenv
requests
